+++ src/main/java/kitchenpos/application/dto/request/MenuProductRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.dto.request;
+package kitchenpos.application.dto.request;
 
 import com.sun.istack.NotNull;
