+++ src/main/java/kitchenpos/application/dto/request/MenuRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.dto.request;
+package kitchenpos.application.dto.request;
 
 import com.sun.istack.NotNull;
 import java.math.BigDecimal;
