+++ src/main/java/kitchenpos/application/dto/request/OrderStatusRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.dto.request;
+package kitchenpos.application.dto.request;
 
 import javax.validation.constraints.NotNull;
 import kitchenpos.domain.order.OrderStatus;
