+++ src/main/java/kitchenpos/ui/ProductRestController.java
@@ -2,9 +2,8 @@ package kitchenpos.ui;
 
 import javax.validation.Valid;
 import kitchenpos.application.ProductService;
-import kitchenpos.dto.request.ProductRequest;
-import kitchenpos.domain.product.Product;
-import kitchenpos.dto.response.ProductResponse;
+import kitchenpos.application.dto.request.ProductRequest;
+import kitchenpos.application.dto.response.ProductResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
